Moved greedy meshing to grid.
